Squashed a couple of warnings.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@1843 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/item.cpp
@@ -174,7 +174,7 @@ bool idItem::UpdateRenderEntity(renderEntity_s *renderEntity, const renderView_t
 			if(inView)
 			{
 				inView = false;
-				lastCycle = ceil(cycle);
+				lastCycle = static_cast<int>(ceil(cycle));
 			}
 		}
